+++ modal-js/examples/sandbox-exec.ts
@@ -1,4 +1,4 @@
-import { App } from "modal";
+import { App, Secret } from "modal";
 
 const app = await App.lookup("libmodal-example", { createIfMissing: true });
 const image = await app.imageFromRegistry("python:3.13-slim");
@@ -35,6 +35,16 @@ for i in range(50000):
     `Got ${contentStdout.length} bytes stdout and ${contentStderr.length} bytes stderr`,
   );
   console.log("Return code:", await p.wait());
+
+  const secret = await Secret.fromName("libmodal-test-secret", {
+    requiredKeys: ["c"],
+  });
+  const printSecret = await sb.exec(["printenv", "c"], {
+    stdout: "pipe",
+    secrets: [secret],
+  });
+  const secretText = await printSecret.stdout.readText();
+  console.log(`Got environment variable c=${secretText}`);
 } finally {
   await sb.terminate();
 }