+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Both client libraries are pre-1.0, and they have separate versioning.
 
 - Added support for snapshotting the filesystem of a Sandbox.
 - Added support for polling Sandboxes to check if they are still running, or get the exit code.
+- Added support to execute commands in Sandboxes with Secrets.
 - Added support for creating Sandboxes with secrets.
 
 ## modal-js/v0.3.14, modal-go/v0.0.14